refactor: move navigation logic to composable functions
Moved the navigation logic for navigating to the SecondScreen and FirstScreen into the composable functions of FirstNavigation and SecondNavigation respectively.
This change removes the need for separate `navigate` functions outside of the composables, making the navigation code more localized and easier to understand.
The logic now directly uses the navController within the composable function's scope, ensuring each screen's navigation is handled in one place.
This also includes setting `launchSingleTop` to true and using `popUpTo` with `inclusive` to clear the backstack when navigating.