feat: add navigation drawer to FirstScreen
Added a navigation drawer to the `FirstScreen`, allowing navigation to other screens.
The `FirstScreen` now uses `AppNavigationDrawer` and `AppScaffold` to display its content and handle drawer interactions.
The `AppNavigationDrawer` has been implemented to accept callbacks for each navigation drawer items.
Added `Third` and `Fourth` to the `Route` for navigate.
The `AppScaffold` composable function now accepts a `modifier` parameter.
The `FirstScreen` now provides functions to navigate to `SecondScreen`, `ThirdScreen` and `FourthScreen`.
Added `FirstScreenClosedPreview` to display preview of the `FirstScreen` in the editor.
Added `DrawerValueParameterProvider` to get values for preview.
Added `LazyColumn` to `FirstScreenContent` as a example.