feat: implement navigation drawer and title icon action
Implemented a navigation drawer within the `AppNavigationDrawer` composable to allow users to navigate between different screens.
The navigation drawer items now close the drawer when clicked using coroutines.
Added a click action to the title icon in `AppScaffold`, which now scrolls the `TweetScreenContent` to the top when clicked.
Added `Route` sealed class `Tweet` property for setup navigation drawer.
Modified `TweetScreen` to use `lazyListState` to scroll the list.
Modified `AppNavHost` to pass the `Route.Tweet` instance to `TweetScreen`.
Added `rememberCoroutineScope` in `TweetScreen` and `AppNavigationDrawer`.
Changed the click action of the items in the `AppNavigationDrawer` to close the navigation drawer after navigation.