chore: remove gradle.xml from .idea
Removed the `gradle.xml` file from the `.idea` directory.
This file is no longer needed for the project's Gradle configuration.

feat: remove .idea directory from repository

Removed the `.idea` directory and its contents from the repository.
The following files were deleted:

-   `.idea/.gitignore`
-   `.idea/gradle.xml`
-   `.idea/misc.xml`
-   `.idea/vcs.xml`

These files are IDE-specific and should not be tracked in the version control system.
This change ensure that the .idea configuration is ignored by the git.